AutoBarcoder/requirements.txt
tkinter
python-levenshtein
networkx
matplotlib
